feat: add deploy button group
